Add runtime compiled version
