+++ pkg/network/ebpf/c/prebuilt/sock-impl.h
@@ -1,5 +1,5 @@
-#ifndef __SOCK_H
-#define __SOCK_H
+#ifndef __SOCK_IMPL_H
+#define __SOCK_IMPL_H
 
 #include "kconfig.h"
 
@@ -271,5 +271,20 @@ static __always_inline int read_conn_tuple(conn_tuple_t* t, struct sock* skp, u6
     return read_conn_tuple_partial(t, skp, pid_tgid, type);
 }
 
+static __always_inline void get_tcp_segment_counts(struct sock* skp, __u32* packets_in, __u32* packets_out) {
+    // counting segments/packets not currently supported on prebuilt
+    // to implement, would need to do the offset-guess on the following
+    // fields in the tcp_sk: packets_in & packets_out (respectively)
+    *packets_in = 0;
+    *packets_out = 0;
+}
 
-#endif
+static __always_inline void* sock_rtt(struct sock *sk) {
+    return (char*)sk + offset_rtt();
+}
+
+static __always_inline void* sock_rtt_var(struct sock *sk) {
+    return (char*)sk + offset_rtt_var();
+}
+
+#endif // __SOCK_IMPL_H
