+++ tasks/system_probe.py
@@ -129,6 +129,7 @@ def test(
     windows=is_windows,
     parallel_build=True,
     failfast=False,
+    kernel_release=None,
 ):
     """
     Run tests on eBPF parts
@@ -148,7 +149,7 @@ def test(
         clang_tidy(ctx)
 
     if not skip_object_files and not windows:
-        build_object_files(ctx, parallel_build=parallel_build)
+        build_object_files(ctx, parallel_build=parallel_build, kernel_release=kernel_release)
 
     build_tags = [NPM_TAG]
     if not windows:
